refactor: gather all server configuration into one file
