+++ .env.example
@@ -1,2 +1,1 @@
-MONGO_URI = 
-SERVER_PORT = 
+MONGO_URI =
